+++ server/src/main/scala/kpn/core/tools/monitor/LongdistanceRouteTool.scala
@@ -105,7 +105,7 @@ class LongdistanceRouteTool(
 
             gpxLineStringOption match {
               case None =>
-                val osmDistance = Math.round(osmRouteSegments.map(_.segment.meters).sum / 1000)
+                val osmDistance = Math.round(osmRouteSegments.map(_.segment.meters).sum.toDouble / 1000)
                 val bounds = Util.mergeBounds(osmRouteSegments.map(_.segment.bounds))
                 val route = LongdistanceRoute(
                   routeRelation.id,
